combine metadata.py and ephemeral.py into one file
